pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wan-video-api"
version = "1.0.0"
description = "Wan Image to Video Generation API"
authors = [{name = "Developer", email = "developer@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "Pillow>=9.5.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
    "xformers>=0.0.22",
    "opencv-python>=4.8.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88